backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
chromadb==0.4.22
PyPDF2==3.0.1
python-docx==1.1.0
zhipuai==2.0.1
pydantic==2.5.3
python-dotenv==1.0.0
numpy==1.26.3
